Refactor Trivy
