+++ ci/go.mod
@@ -1,0 +1,8 @@
+module github.com/fairwindsops/insights-plugins/ci
+
+go 1.13
+
+require (
+	github.com/sirupsen/logrus v1.6.0
+	github.com/thoas/go-funk v0.6.0
+)
