+++ trivy/go.mod
@@ -6,8 +6,9 @@ require (
 	github.com/Azure/go-autorest v11.1.2+incompatible // indirect
 	github.com/sirupsen/logrus v1.6.0
 	github.com/thoas/go-funk v0.6.0
-	k8s.io/api v0.17.4
-	k8s.io/apimachinery v0.17.3
-	k8s.io/client-go v0.17.4
+	k8s.io/api v0.18.2
+	k8s.io/apimachinery v0.18.2
+	k8s.io/client-go v0.18.2
 	sigs.k8s.io/controller-runtime v0.6.0
+	sigs.k8s.io/structured-merge-diff v0.0.0-20190525122527-15d366b2352e // indirect
 )
